requirements.txt
asgiref==3.7.2
attrs==23.1.0
build==0.10.0
click==8.1.6
cmake==3.21.4
colorama==0.4.6
Django==4.2.4
django-restframework==0.0.1
djangorestframework==3.14.0
dlib @ file:///C:/Users/naren/Downloads/dlib-19.22.99-cp39-cp39-win_amd64.whl
drf-spectacular==0.26.4
face-recognition==1.3.0
face-recognition-models==0.3.0
inflection==0.5.1
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pyproject_hooks==1.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.9.2
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
